+++ Sources/EpoxyBars/BarModel/BarModelBuilder.swift
@@ -1,0 +1,7 @@
+// Created by eric_horacek on 3/15/21.
+// Copyright © 2021 Airbnb Inc. All rights reserved.
+
+import EpoxyCore
+
+/// A result builder that enables a DSL for building arrays of bar models.
+public typealias BarModelBuilder = EpoxyModelArrayBuilder<BarModeling>