+++ Sources/EpoxyBars/BottomBarInstaller/BottomBarInstaller.swift
@@ -30,6 +30,14 @@ public final class BottomBarInstaller: NSObject {
     installer.setBars(bars, animated: false)
   }
 
+  public convenience init(
+    viewController: UIViewController,
+    avoidsKeyboard: Bool = false,
+    @BarModelBuilder bars: () -> [BarModeling])
+  {
+    self.init(viewController: viewController, bars: bars())
+  }
+
   // MARK: Public
 
   /// The container installed in the view controller's view that contains the bar stack.